requirements.txt
absl-py==2.1.0
accelerate==0.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.6.2.post1
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.2.0
av==14.1.0
bitarray==3.3.1
blis==1.0.1
boto3==1.35.69
botocore==1.35.69
braceexpand==0.1.7
bravado==11.0.3
bravado-core==6.1.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.0
click==8.1.7
clip
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.10
Cython==3.0.12
decorator==5.1.1
decord==0.6.0
deepspeed==0.17.0
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
dukpy==0.5.0
easydict==1.13
einops==0.3.0
exceptiongroup==1.2.2
fairscale==0.4.4
fairseq==0.12.2
fastapi==0.115.5
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.16.1
flatten-dict==0.4.2
fonttools==4.55.0
frozenlist==1.5.0
fschat==0.1.10
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==3.23.0
greenlet==3.2.1
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
hydra-core==1.0.7
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.6.1
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpickle==1.5.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==3.2.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.77.0
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
lxml==5.3.2
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mne==1.8.0
monai==1.4.0
monotonic==1.6
moviepy==2.0.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munch==2.5.0
murmurhash==1.0.11
narwhals==1.14.2
neptune-client==0.16.7
neptune-sacred==0.10.0
networkx==3.2.1
ninja==1.11.1.4
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.575.51
oauthlib==3.2.2
omegaconf==2.0.5
openai==0.28.0
opencv-python==4.10.0.84
optuna==4.3.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathtools==0.1.2
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.1
psutil==6.1.0
py-cpuinfo==9.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.0
pynndescent==0.5.13
pyparsing==3.2.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytube==15.0.0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
s3transfer==0.10.4
sacrebleu==2.5.1
sacred==0.8.2
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.4.1
sentencepiece==0.1.97
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
simplejson==3.18.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
socks==0
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.4.8
starlette==0.41.3
submitit==1.5.2
svgwrite==1.4.3
swagger-spec-validator==3.0.4
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.5.1
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.4.12
tokenizers==0.15.2
torch 
torchaudio 
torchinfo==1.8.0
torchvision 
tqdm==4.65.2
transformers==4.39.3
triton==2.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
umap-learn==0.5.7
urllib3==1.26.20
uvicorn==0.32.1
wandb==0.14.2
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdataset==0.1.103
websocket-client==1.8.0
websockets==14.1
Werkzeug==2.2.2
wrapt==1.17.0
yacs==0.1.8
yarl==1.18.0
you-get==0.4.1730
youtube_dl 
yt-dlp==2025.1.26
zipp==3.21.0
